refactor: migrate check ingang apply period to middleware
